Simplify the DAG, only use a single, global DAG for dependencies.
Previously, this followed the approach in Build System Rules and
Algorithms: http://gittup.org/tup/build_system_rules_and_algorithms.pdf

Where a "Partial DAG" is created incrementally upon processing a change.
This means writes would traverse the global DAG on each write.

Now, we simplify the traversal of the DAG such that nodes in the DAG are
marked as dirty. When visiting topographically, we only visit the
subgraph reachable by the known dirty nodes. When visited, a dirty node
is "processed" (its data recalculated) -- if it is deemed dirty (it
returns false), then we mark all the adjacent destination nodes as
dirty. If it is not deemed dirty, we pass through.

This allows for a simpler early-exit of processing changes in the graph
(no need for refcounting anymore) and also allows for inert nodes to
live in the graph: they only exist to specify topographic ordering.