Guarantee the ordering of onEffect() calls to be after mounted calc() calls
This change introduces a big change into the DAG. There are two flavors of
edges in the DAG: HARD and SOFT. Internally, an edge may be tagged with
both HARD and SOFT, in which case it is treated as HARD.

* HARD edges represent data and order dependencies
* SOFT edges represent order dependencies

When a node is dirty, visited, and propagation continues, its HARD edges
will be visited next. Its SOFT edges are not visited and only are used
for topographic sorting of visits.

In the view layer, each calc() mounted node introduces a new kind of
node in the DAG, which solely represents ordering. Soft edges are
created between child -> parent ordering nodes; a component's onEffect
has a soft edge between the parent calc() -> onEffect. This ensures that
onEffects are executed after their containing calc()s are all performed.

Because of the view layer changes, we no longer need to manually
retain() calc()s, and now only retain the root ordering node.