Only mark nodes as dirty if they have hard dependencies
Additionally, add some extra debug code (debugSubscribe) that allows for
the incremental display of the graph as it is garbage collected &
visited topographically. And some helper code to build a lightweight UI
around displaying this graph.

As a performance optimization, if a node should be come dirty, it only
really matters if it has a data dependency: if some other node cares
about changes in the node.

This allows for skipping a large number of visits in the DAG, especially
when subscriptions are involved.

To accomplish this, each observed model has a main node and a
subscription node. Edges are ordered as such:
* The main node has a hard edge to the subscription node
* The main node has a soft edge to all the fields
* All the fields have a soft edge to the subscription node

When a field becomes modified:
* The field node is dirtied
* The subscription node is dirtied

When the subscription node is visited
* The subscription node "pushes" change events to the observing (main) nodes

When the observing (main) nodes are visited:
* The main node processes these events, potentially dirtying its fields

This ensures proper ordering of dependencies, and ensures that
unobserved nodes do not have this additional overhead.