Adjust recalculation of DAG to account for mutating DAG
This is certainly not "correct" but this change allows for the
adjustment of the DAG while processing to "work as one would expect"
in that the following are true:

* Dirty nodes that do not reach retained nodes, but change while the DAG
  is processed so they do reach retained nodes are processed
* View root nodes that dirty _new_ nodes while being processed which
  reach retained nodes are processed.

Note: this completely changes how garbage collection is done. Now, lots
of garbage is left behind. Eventually, a mark+sweep garbage collection
should be added, which should be performed periodically.