Pre-sort nodes to reduce bulk insert work in common cases
Bulk edge additions/removals occur when _new nodes_ are added to the
graph. To prevent the particularly bad case where we add nodes in the
incorrect order, then add edges, which cause resorting of the nodes we
just added; we pre-sort the nodes in (partial) topological order based
on the edges also being added.

This _vastly_ reduces the amount of time taken (and the number of
resorts performed) when mounting new components that contain state.

When inserting 10k nodes via the benchmark demo
- Without pre-sorting: 90,004 reorder() calls, sorting a total of 410,020 items
- With pre-sorting: 5 reorder() calls, sorting a total of 16 items