Bugfix: ensure graph is fully processed when complete
It used to be the case that after .process() finished, the graph could
contain dirty vertices. This could happen if a vertex at topological
index N were to mark a vertex at topological index < N as dirty as a
result of processing.

Now we maintain a low index of the dirty vertices and continue from
there.

Also reorders all symbols into a single symbols file for easier importing.