Remove Contexts from Gooey
It adds far more complexity than I think it's worth. If it needs to be
brought back this commit can always be referenced.

Specifically, it makes reasoning about the lifecycles of Components much
harder, since context values may change after the component itself is
retained. Perhaps there's another way of solving the "context" problem,
but connecting it to the data model, so that reading a context feels
like reading a calculation -- so any dependent calculations will
recalculate themselves when the context has changed...

But that's a big change for questionable benefit.