Fix two derived view auto-retaining and model.keys
Previously, you had to manually retain a derived view for it to be
usable within a calculation. Now if a derived view is accessed within a
calculation, the calculation retains the derived view. As it should!

Previously, model.keys was totally broken. It only worked via
subscription, not via invalidation of its fields. Now it correctly
invalidates its keys so it can be used in a calculation.