Minor internal improvement: avoid tracking SymProcessable in TrackedData
The SymProcessable internal symbol indicates whether or not a dependency
should be automatically tracked when read / accessed. This is only true
for Processable interfaces (calculations, fields, subscription
consumers, and subscription emitters.

TrackedData objects were _not_ processable, but this field access was
being looked up on the object. This avoids a Reflect lookup for these
proxies and adds debug assertions to ensure additional internal symbols
aren't unintentionally looked up.

The lookup removal may not actually be needed since we do not track
symbol accesses, but the assertions are a good idea.