Fix subtle bug in graph processing, rely only on one adjacency list
This change reduces memory and processing time maintaining two sets of
adjacency lists: forward adjacency and reverse adjacency.

Additionally, there was a subtle race condition where if a vertex was
removed from a cycle, vertex that depended on that cycle would not
necessarily be marked as dirty. Now we dirty all vertexes that depend on
a cycle when that cycle changes members (gains or removes a vertex).