Batch RenderNode DOM removals and inserts
This is a very big change.

Previously, when a RenderNode responded to a calculation / collection
changing, the DOM removals would happen first, and then insertions would
happen at the end of the flush. Due to how these events were triggered,
they were not batched, and insertions to the same parent would be
performed in multiple parentNode.insertBefore() calls.

Now, we hold a second directed graph (hey, we have one) which holds the
(virtual) DOM structure, which is dirtied when nodes need to be mutated
/ mount / unmount. This directed graph is then pre-processed, node removals
performed in batch, node insertions performed in batch, and
post-processed in the correct topological order. This batching allows
node insertions and removals to be grouped together.