Bugfix: infinite commit loop in certain cases
Previously, IntrinsicRenderNodes were not correctly reconciling child
nodes, causing unnecessarily moved nodes from being inserted into a
fragment, or the same parent via .insertBefore. This causes the user
agent to lose focus, and trigger a 'blur' event.

Note: there still is an awkward usage of insertBefore, where we rely on
that to reorder nodes in response to a .moveSlice event.