NEW FEATURE: custom elements
Wow, this worked *way* better than I was expecting.

To define a custom autonomous (non-built-in) element, use:

    defineCustomElement({
        tagName: 'my-greeting',
        attributes: ['name', 'excitement'],
        Component: ({ name, excitement, children }) => {
            return (
                <>Hello, {name}{calc(() => {
                    const level = parseInt(dynGet(excitement), 10);
                    return isNaN(level) ? null : '!' * level;
                })}</>
            );
        },
    });

    // Optional: Register JSX types for this element
    declare module "@srhazi/gooey" {
        interface CustomElements {
            'my-greeting': {
                name?: Dyn<string | undefined>;
                excitement?: Dyn<string | undefined>;
            }
        }
    }

This element is available in HTML as <my-greeting name="world" excitement="10"></my-greeting>

To define a custom element that extends from a built-in element, use:

    defineCustomElement({
        tagName: 'my-greeting',
        attributes: ['name', 'excitement'],
        extends: 'div',
        Component: ({ name, excitement, children }) => {
            return (
                <>Hello, {name}{calc(() => {
                    const level = parseInt(dynGet(excitement), 10);
                    return isNaN(level) ? null : '!' * level;
                })}</>
            );
        },
    });

This element is available in HTML as <div is="my-greeting" name="world" excitement="10"></div>.
NOTE: this is not implemented in Safari (https://bugs.webkit.org/show_bug.cgi?id=182671)

The shadowMode property can be set to 'open' or 'closed', in which case
the custom element renders its contents to its shadow root. There is no
children property passed to the component, and additional methods