FEATURE: provide a solution around HTML parsing limitations
HTML parsing is complicated, and when hydrating web components this can
cause some serious surprises.

For example, the following HTML will be hydrated into DOM trees as the
following:

HTML:
    <div id="parent">
        <p id="without-template">
            Before-text
            <custom-element>
                <div slot="foo">slot-foo</div>
                <div slot="bar">slot-bar</div>
            </custom-element>
            After-text
        </p>
    </div>

DOM tree (purely whitespace Text nodes omitted):
- DIV#parent
    - P#without-template
      - Text: Before-text
      - CUSTOM-ELEMENT (empty, only whitespace Text children)
    - DIV[slot="foo"]
      - Text: slot-foo
    - DIV[slot="foo"]
      - Text: slot-bar
    - Text: After text
    - P (empty, no children)

This is because <p> elements cannot have child <div> elements when being constructed via parsing html.

One potential way to get around the parsing limitations is to use <template>
elements, which create a different parsing context and can live within
any element.

So we allow (by default, with opt-out via the `hydrateTemplateChild:
false` option) custom elements to have a single <template> child
element, which is replaced with its contents DocumentFragment upon
initializing & rendering the custom element.

This comes at a cost of limiting Graceful Degradation, as <template>
elements are rendered completely inert as nothing. So if there is a
failure in loading / defining the custom element, the "fallback" content
of the custom element will be blank. This may be an appropriate
fallback.