Adjust the RenderNode interface for performance
Previously, .detach() was meant to tell the RenderNode instance to
detach itself. This is inefficient, as when you remove a RenderNode
child, each of the children would need to detach themselves.

Since each RenderNode can keep track of how many nodes have been emitted
by its child, .detach() can purely be a notification mechanism: you have
been detached. And the RenderNode that is detaching one of its children
can emit a splice event that removes all of the Nodes emitted.