Big change: change graph propagation & models/collections graph representation
This change rewrites almost the *entire* model layer. Calculations,
collections, views, and dicts are all impacted.

Before this change, it was the responsibility of the graph to propagate
changes downward, following the direction of "hard" edges in the graph
and using the presence of these "hard" edges to identify the nodes to
propagate dirtiness.

This was a problem for maintaining topology in collections. See
https://abstract.properties/gooey-graph-internals.html for why this was
not great.

tl;dr: because collections & views had fine-grained nodes in the graph,
inserting a new node to a collection would require a full topological
sort of its entire mapped chain. For large collections this was slow.

This change moves the identification of nodes to propagate from the
structure of the dependency graph (hard edges) to the responsibility of
the node being processed.

Naive nodes can ask the graph for their downstream dependencies and use
that information.

Smart nodes (like collections/views/dicts/models) are now able to know
which calculation is reading them, so they can hold an internal mapping
between which calculation depends on which slice of data.

This collapses a collection of 100 items & mapped view from 204 nodes
(100 collection nodes, 1 collection length node, 1 collection emitter, 1
view consumer, 100 view nodes, 1 view length node) to 2 nodes in the
graph. Nice!

---

This change **also** adjusts some interfaces slightly and fixes bugs.
Here are the changes:

- BUGFIX: Fixed very subtle infinite loop / hanging cycle bug
- BREAKING: Dict#keys() renamed to Dict#keysView()
- BREAKING: (minor) model.subscribe() synchronously logs current model fields
- BREAKING: (minor) model.field() returns a DynamicMut, not a Field