Version 0.21.0
A note about this version:

This version includes a *major* rewrite of the data layer. It
significantly improves the performance of changing items in a
collection.

Gooey is approaching version 1.0, which I consider to be the "done"
version. That is to say, the version at which major breaking changes are
not performed, and development is solely focused on bugfixes,
performance optimizations, and exposing engine internals for better
debugging tools.

Changes since 0.20.0:

- BUGFIX: Fixed very subtle infinite loop / hanging cycle bug
- BREAKING: Dict has adjusted some methods. .entries(), .keys(), and
  .values() return plain old arrays. A new .keysView() returns a view of
  the keys.
- BREAKING: model.subscribe(), collection.subscribe(), and
  dict.subscribe() all synchronously call the callback on first
  invocation with their current fields, allowing subscribers to know the
  contents of the model/collection/dict without having to reach around
  and read the contents before subscribing.
- BREAKING: model.field() returns a DynamicMut, not a Field