Improve performance of dirtying ArraySub during insert/delete
Previously, an insert/delete would need to invalidate each individual index from the
insertion point to the end of the array.

This introduces a new "RangeAssociation" abstraction that allows us to
associate *ranges* in array with a value, which is the change clock.

Instead of O(n) operations for big arrays, it's now O(log(n)).

NOTE: this adds duplication between -- trackeddata.ts and trackedarray.ts

Need to follow up with a refactor to lift out common functionality